Refactor: Move clustering_regime_detection to nas_tas utils
